notification-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
kafka-python==2.0.2
python-json-logger==2.0.7